+++ spring-ms-common-api/src/main/java/com/geekbrains/springms/api/OrderItemDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +14,7 @@ import lombok.Setter;
 public class OrderItemDto {
     private Long id;
     private Long productId;
-    private Double price;
+    private BigDecimal price;
     private Integer amount;
-    private Double sum;
+    private BigDecimal sum;
 }
